Update change theme logix to remove theme flickering issue
Because nextjs does not know what is happening on the client side, the
change theme logic was causing a flickering issue when the user changed
the page. Now it stopped by adding a stopper until the theme is decided
on the client side.

**note**: Unfortunately, the fix only work on production which caused me
to add a workaround to make it work on development. But it needs to be
addressed in the future.