Update styling of the mdx components
Now there is a new type of Callout, the `informative`, and the dark mode
was fixed.

Also the behavior of the title link changed, before we had it underline,
now it adds a small hashtag at the end of it indicating that we can click
and get the link to that title.